+++ mempool/mempool.go
@@ -29,7 +29,12 @@ import (
 
 var _ sdkmempool.ExtMempool = &ExperimentalEVMMempool{}
 
-const SubscriberName = "evm"
+const (
+	// SubscriberName is the name of the event bus subscriber for the EVM mempool
+	SubscriberName = "evm"
+	// fallbackBlockGasLimit is the default block gas limit is 0 or missing in genesis file
+	fallbackBlockGasLimit = 100_000_000
+)
 
 type (
 	// ExperimentalEVMMempool is a unified mempool that manages both EVM and Cosmos SDK transactions.
@@ -52,6 +57,7 @@ type (
 		bondDenom     string
 		evmDenom      string
 		blockGasLimit uint64 // Block gas limit from consensus parameters
+		minTip        *uint256.Int
 
 		/** Verification **/
 		anteHandler sdk.AnteHandler
@@ -72,6 +78,7 @@ type EVMMempoolConfig struct {
 	AnteHandler      sdk.AnteHandler
 	BroadCastTxFn    func(txs []*ethtypes.Transaction) error
 	BlockGasLimit    uint64 // Block gas limit from consensus parameters
+	MinTip           *uint256.Int
 }
 
 // NewExperimentalEVMMempool creates a new unified mempool for EVM and Cosmos transactions.
@@ -80,10 +87,8 @@ type EVMMempoolConfig struct {
 // of pools and verification functions, with sensible defaults created if not provided.
 func NewExperimentalEVMMempool(getCtxCallback func(height int64, prove bool) (sdk.Context, error), logger log.Logger, vmKeeper VMKeeperI, feeMarketKeeper FeeMarketKeeperI, txConfig client.TxConfig, clientCtx client.Context, config *EVMMempoolConfig) *ExperimentalEVMMempool {
 	var (
-		txPool      *txpool.TxPool
-		cosmosPool  sdkmempool.ExtMempool
-		anteHandler sdk.AnteHandler
-		blockchain  *Blockchain
+		cosmosPool sdkmempool.ExtMempool
+		blockchain *Blockchain
 	)
 
 	bondDenom := evmtypes.GetEVMCoinDenom()
@@ -98,13 +103,12 @@ func NewExperimentalEVMMempool(getCtxCallback func(height int64, prove bool) (sd
 		panic("config must not be nil")
 	}
 
-	anteHandler = config.AnteHandler
-	blockchain = newBlockchain(getCtxCallback, logger, vmKeeper, feeMarketKeeper, config.BlockGasLimit)
-
 	if config.BlockGasLimit == 0 {
-		logger.Debug("block gas limit is 0, setting default", "default_limit", 100_000_000)
-		config.BlockGasLimit = 100_000_000
+		logger.Warn("block gas limit is 0, setting to fallback", "fallback_limit", fallbackBlockGasLimit)
+		config.BlockGasLimit = fallbackBlockGasLimit
 	}
+
+	blockchain = newBlockchain(getCtxCallback, logger, vmKeeper, feeMarketKeeper, config.BlockGasLimit)
 
 	// Create txPool from configuration
 	legacyConfig := legacypool.DefaultConfig
@@ -180,7 +184,8 @@ func NewExperimentalEVMMempool(getCtxCallback func(height int64, prove bool) (sd
 		bondDenom:     bondDenom,
 		evmDenom:      evmDenom,
 		blockGasLimit: config.BlockGasLimit,
-		anteHandler:   anteHandler,
+		minTip:        config.MinTip,
+		anteHandler:   config.AnteHandler,
 	}
 
 	vmKeeper.SetEvmMempool(evmMempool)
@@ -448,7 +453,7 @@ func (m *ExperimentalEVMMempool) getIterators(goCtx context.Context, i [][]byte)
 	m.logger.Debug("getting iterators")
 
 	pendingFilter := txpool.PendingFilter{
-		MinTip:       nil,
+		MinTip:       m.minTip,
 		BaseFee:      baseFeeUint,
 		BlobFee:      nil,
 		OnlyPlainTxs: true,