+++ server/config/toml.go
@@ -22,6 +22,9 @@ cache-preimage = {{ .EVM.EnablePreimageRecording }}
 # EVMChainID is the EIP-155 compatible replay protection chain ID. This is separate from the Cosmos chain ID.
 evm-chain-id = {{ .EVM.EVMChainID }}
 
+# MinTip defines the minimum priority fee for the mempool.
+min-tip = {{ .EVM.MinTip }}
+
 ###############################################################################
 ###                           JSON RPC Configuration                        ###
 ###############################################################################